Convert to input only in FlowDSL
